Move the django_models module to django.models #48
- Deprecation warning included to provide hint on the migration
